refactor: RacingCarConfig 클래스의 이름을 Constant로 변경
